Detailed-Planning-Modified

+++ CWB.App/Models/BusinessProcesses/ChildWoRelVM.cs
@@ -10,7 +10,7 @@ namespace CWB.App.Models.BusinessProcesses
         public long ChildWoRelId { get; set; }
         public long WoId { get; set; }
         public long PartId { get; set; }
-        public int Qnty { get; set; }
+        public decimal Qnty { get; set; }
         public string CameFrom { get; set; }
         public long TenantId { get; set; }
     }
